pyproject.toml
[tool.poetry]
name = "jarvis"
version = "1.0.0"
description = "Voice-Driven Tool-Orchestrated Automation System"
authors = ["X0ditya"]
readme = "README.md"
packages = [{include = "audio"}, {include = "stt"}, {include = "commands"}, {include = "core"}, {include = "security"}, {include = "infra"}]


[tool.poetry.dependencies]
python = "^3.10"
sounddevice = "^0.4.6"
numpy = ">=1.24.0"
faster-whisper = "^1.0.0"
pyyaml = "^6.0"
pydantic = "^2.0.0"
pynput = "^1.7.6"
rich = "^13.0.0"
# Phase 2 dependencies
openai = "^1.0.0"
httpx = "^0.25.0"
# Gemini support
google-generativeai = "^0.8.0"
# Infrastructure dependencies
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
python-multipart = "^0.0.12"
# TTS dependencies
piper-tts = { version = "^1.2.0", optional = true }
edge-tts = "^6.1.0"
# Multimodal/Camera (optional)
opencv-python = { version = "^4.8.0", optional = true }
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
ruff = "^0.1.0"

[tool.poetry.scripts]
jarvis = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
